requirements.docker.txt
Flask
gunicorn
python-dotenv

psycopg2-binary
boto3
requests

numpy
pillow
opencv-python-headless

ultralytics

easyocr
pyclipper
python-bidi
shapely
timm
einops
pytorch-lightning
omegaconf
hydra-core
